Break up model tests into their own folder
