+++ tests/unit/models/test_consultation.py
@@ -1,0 +1,12 @@
+import pytest
+
+from django.db import IntegrityError
+
+from consultation_analyser import factories
+
+
+@pytest.mark.django_db
+def test_uniqueness_consultation_slugs():
+    factories.ConsultationFactory(name="My new consultation", slug="my-new-consultation")
+    with pytest.raises(IntegrityError):
+        factories.ConsultationFactory(name="My new consultation 2", slug="my-new-consultation")
